requirements.txt
azure-functions==1.8.0
azure-functions-durable==1.1.3